Revert "feat: move flatten type to shared folder"
This reverts commit 517382c47a044ae58b1f9e6a4c4a49e95b68071b.
