+++ packages/ui/react/src/colors/index.ts
@@ -1,1 +1,2 @@
 export * from './getColor'
+export * from './types'
